Cleaned up plenty of comments. Put some information to variables (i.e. functions.php, bookmark id, url, displayname, & visits) for mainenance. Cleared up some headers since you can just make the end of the function cause a header (i.e. bookmarks.php, adding a bookmark can just set either message, and exit right at the end.) Deleted previous-styles.
I was able to finally fix the visit counter by setting the curl_init to the URL of the web services. For some reason the browser never caught that before. I noticed that the bookmarks are preemptively saved in order by visits, so when they get displayed they are already in order. The final project is essentially finished, with all of our requirements met.